<==== Mercurial data ====>
Branch: vms_4.0_server

+++ vms/libs/common/src/nx/vms/common/p2p/downloader/private/storage.cpp
@@ -13,15 +13,41 @@
 
 using namespace std::chrono;
 
+namespace nx::vms::common::p2p::downloader {
+
 namespace {
 
 static constexpr qint64 kDefaultChunkSize = 1024 * 1024;
 static const QString kMetadataSuffix = ".vmsdownload";
 static const milliseconds kCleanupPeriod = 5min;
 
+static FileMetadata prepareInitialFileMetadata(const FileInformation& fileInformation,
+    const QDir& defaultDownloadsDirectory)
+{
+    FileMetadata info = fileInformation;
+    if (info.chunkSize <= 0)
+        info.chunkSize = kDefaultChunkSize;
+
+    if (info.absoluteDirectoryPath.isEmpty())
+    {
+        info.fullFilePath = defaultDownloadsDirectory.absoluteFilePath(info.name);
+        info.absoluteDirectoryPath = QFileInfo(info.fullFilePath).absolutePath();
+    }
+    else
+    {
+        info.fullFilePath = QDir(info.absoluteDirectoryPath).absoluteFilePath(info.name);
+    }
+
+    return info;
+}
+
+static QString keyFromFileName(const QString& fileName)
+{
+    return QFileInfo(fileName).fileName();
+}
+
 } // namespace
 
-namespace nx::vms::common::p2p::downloader {
 
 QN_FUSION_DECLARE_FUNCTIONS(FileMetadata, (json))
 
@@ -45,22 +71,31 @@ QDir Storage::downloadsDirectory() const
     return m_downloadsDirectory;
 }
 
-QStringList Storage::files() const
-{
-    QnMutexLocker lock(&m_mutex);
-    return m_fileInformationByName.keys();
-}
-
 QString Storage::filePath(const QString& fileName) const
 {
-    return m_downloadsDirectory.absoluteFilePath(fileName);
+    QnMutexLocker lock(&m_mutex);
+    const auto it = m_fileInformationByName.find(keyFromFileName(fileName));
+    if (it == m_fileInformationByName.cend())
+        return QString();
+
+    return it.value().fullFilePath;
 }
 
-FileInformation Storage::fileInformation(
-    const QString& fileName) const
+QStringList Storage::files() const
+{
+    QStringList result;
+    {
+        QnMutexLocker lock(&m_mutex);
+        for (auto it = m_fileInformationByName.cbegin(); it != m_fileInformationByName.cend(); ++it)
+            result.append(it.value().fullFilePath);
+    }
+    return result;
+}
+
+FileInformation Storage::fileInformation(const QString& fileName) const
 {
     QnMutexLocker lock(&m_mutex);
-    return m_fileInformationByName.value(fileName);
+    return m_fileInformationByName.value(keyFromFileName(fileName));
 }
 
 ResultCode Storage::addFile(FileInformation fileInformation, bool updateTouchTime)
@@ -80,21 +115,14 @@ ResultCode Storage::addDownloadedFile(const FileInformation& fileInformation)
 
     QnMutexLocker lock(&m_mutex);
 
-    if (m_fileInformationByName.contains(fileInformation.name))
+    if (m_fileInformationByName.contains(fileInformation.key()))
         return ResultCode::fileAlreadyExists;
 
-    FileMetadata info = fileInformation;
-    const auto path = filePath(fileInformation.name);
-
-    if (info.chunkSize <= 0)
-        info.chunkSize = kDefaultChunkSize;
-
-    QFile file(path);
-
-    if (!file.exists())
+    FileMetadata info = prepareInitialFileMetadata(fileInformation, m_downloadsDirectory);
+    if (!QFile(info.fullFilePath).exists())
         return ResultCode::fileDoesNotExist;
 
-    const auto md5 = calculateMd5(path);
+    const auto md5 = calculateMd5(info.fullFilePath);
     if (md5.isEmpty())
         return ResultCode::ioError;
 
@@ -103,7 +131,7 @@ ResultCode Storage::addDownloadedFile(const FileInformation& fileInformation)
     else if (info.md5 != md5)
         return ResultCode::invalidChecksum;
 
-    const auto size = calculateFileSize(path);
+    const auto size = calculateFileSize(info.fullFilePath);
     if (size < 0)
         return ResultCode::ioError;
 
@@ -113,7 +141,7 @@ ResultCode Storage::addDownloadedFile(const FileInformation& fileInformation)
         return ResultCode::invalidFileSize;
 
     const int chunkCount = calculateChunkCount(info.size, info.chunkSize);
-    info.chunkChecksums = calculateChecksums(path, info.chunkSize);
+    info.chunkChecksums = calculateChecksums(info.fullFilePath, info.chunkSize);
     if (info.chunkChecksums.size() != chunkCount)
         return ResultCode::ioError;
 
@@ -122,7 +150,7 @@ ResultCode Storage::addDownloadedFile(const FileInformation& fileInformation)
     if (!saveMetadata(info))
         return ResultCode::ioError;
 
-    m_fileInformationByName.insert(fileInformation.name, info);
+    m_fileInformationByName.insert(fileInformation.key(), info);
 
     lock.unlock();
 
@@ -137,31 +165,25 @@ ResultCode Storage::addNewFile(const FileInformation& fileInformation)
 
     QnMutexLocker lock(&m_mutex);
 
-    if (m_fileInformationByName.contains(fileInformation.name))
+    if (m_fileInformationByName.contains(fileInformation.key()))
         return ResultCode::fileAlreadyExists;
 
-    FileMetadata info = fileInformation;
-    const auto path = filePath(fileInformation.name);
-
-    if (info.chunkSize <= 0)
-        info.chunkSize = kDefaultChunkSize;
-
-    const auto fileDir = QFileInfo(path).absolutePath();
-    if (!QDir(fileDir).exists())
+    FileMetadata info = prepareInitialFileMetadata(fileInformation, m_downloadsDirectory);
+    if (!QDir(info.absoluteDirectoryPath).exists())
     {
-        if (!QDir().mkpath(fileDir))
+        if (!QDir().mkpath(info.absoluteDirectoryPath))
             return ResultCode::ioError;
     }
 
-    if (!info.md5.isEmpty() && calculateMd5(path) == info.md5)
+    if (!info.md5.isEmpty() && calculateMd5(info.fullFilePath) == info.md5)
     {
         info.status = FileInformation::Status::downloaded;
-        info.size = calculateFileSize(path);
+        info.size = calculateFileSize(info.fullFilePath);
         if (info.size < 0)
             return ResultCode::ioError;
 
         const int chunkCount = calculateChunkCount(info.size, info.chunkSize);
-        info.chunkChecksums = calculateChecksums(path, info.chunkSize);
+        info.chunkChecksums = calculateChecksums(info.fullFilePath, info.chunkSize);
         if (info.chunkChecksums.size() != chunkCount)
             return ResultCode::ioError;
 
@@ -188,7 +210,7 @@ ResultCode Storage::addNewFile(const FileInformation& fileInformation)
             info.chunkChecksums.resize(chunkCount);
         }
 
-        ResultCode reserveStatus = reserveSpace(path, info.size >= 0 ? info.size : 0);
+        ResultCode reserveStatus = reserveSpace(info.fullFilePath, info.size >= 0 ? info.size : 0);
         if (reserveStatus != ResultCode::ok)
             return reserveStatus;
 
@@ -198,7 +220,7 @@ ResultCode Storage::addNewFile(const FileInformation& fileInformation)
     if (!saveMetadata(info))
         return ResultCode::ioError;
 
-    m_fileInformationByName.insert(fileInformation.name, info);
+    m_fileInformationByName.insert(fileInformation.key(), info);
 
     lock.unlock();
 
@@ -207,12 +229,12 @@ ResultCode Storage::addNewFile(const FileInformation& fileInformation)
     return ResultCode::ok;
 }
 
-ResultCode Storage::updateFileInformation(
-    const QString& fileName, qint64 size, const QByteArray& md5)
+ResultCode Storage::updateFileInformation(const QString& fileName, qint64 size,
+    const QByteArray& md5)
 {
     QnMutexLocker lock(&m_mutex);
 
-    auto it = m_fileInformationByName.find(fileName);
+    auto it = m_fileInformationByName.find(keyFromFileName(fileName));
     if (it == m_fileInformationByName.end())
         return ResultCode::fileDoesNotExist;
 
@@ -229,7 +251,7 @@ ResultCode Storage::updateFileInformation(
         const int chunkCount = calculateChunkCount(size, it->chunkSize);
         it->downloadedChunks.resize(chunkCount);
         it->chunkChecksums.resize(chunkCount);
-        resizeFailed = !QFile::resize(filePath(fileName), size);
+        resizeFailed = !QFile::resize(it->fullFilePath, size);
 
         updated = true;
     }
@@ -272,7 +294,7 @@ ResultCode Storage::setChunkSize(const QString& fileName, qint64 chunkSize)
 
     QnMutexLocker lock(&m_mutex);
 
-    auto it = m_fileInformationByName.find(fileName);
+    auto it = m_fileInformationByName.find(keyFromFileName(fileName));
     if (it == m_fileInformationByName.end())
         return ResultCode::fileDoesNotExist;
 
@@ -307,12 +329,11 @@ ResultCode Storage::setChunkSize(const QString& fileName, qint64 chunkSize)
     return ResultCode::ok;
 }
 
-ResultCode Storage::readFileChunk(
-    const QString& fileName, int chunkIndex, QByteArray& buffer)
+ResultCode Storage::readFileChunk(const QString& fileName, int chunkIndex, QByteArray& buffer)
 {
     QnMutexLocker lock(&m_mutex);
 
-    auto it = m_fileInformationByName.find(fileName);
+    auto it = m_fileInformationByName.find(keyFromFileName(fileName));
     if (it == m_fileInformationByName.end())
         return ResultCode::fileDoesNotExist;
 
@@ -322,7 +343,7 @@ ResultCode Storage::readFileChunk(
     if (!it->downloadedChunks.testBit(chunkIndex))
         return ResultCode::invalidChunkIndex;
 
-    QFile file(filePath(it->name));
+    QFile file(it->fullFilePath);
     if (!file.open(QFile::ReadOnly))
         return ResultCode::ioError;
 
@@ -337,12 +358,11 @@ ResultCode Storage::readFileChunk(
     return ResultCode::ok;
 }
 
-ResultCode Storage::writeFileChunk(
-    const QString& fileName, int chunkIndex, const QByteArray& buffer)
+ResultCode Storage::writeFileChunk(const QString& fileName, int chunkIndex, const QByteArray& buffer)
 {
     QnMutexLocker lock(&m_mutex);
 
-    auto it = m_fileInformationByName.find(fileName);
+    auto it = m_fileInformationByName.find(keyFromFileName(fileName));
     if (it == m_fileInformationByName.end())
         return ResultCode::fileDoesNotExist;
 
@@ -352,7 +372,7 @@ ResultCode Storage::writeFileChunk(
     if (it->status == FileInformation::Status::downloaded)
         return ResultCode::ioError;
 
-    QFile file(filePath(it->name));
+    QFile file(it->fullFilePath);
     if (!file.open(QFile::ReadWrite)) //< ReadWrite because WriteOnly implies Truncate.
         return ResultCode::ioError;
 
@@ -400,47 +420,49 @@ ResultCode Storage::deleteFile(const QString& fileName, bool deleteData)
             return resultCode;
     }
 
-    emit fileDeleted(fileName);
+    emit fileDeleted(filePath(fileName));
     return ResultCode::ok;
 }
 
 ResultCode Storage::deleteFileInternal(const QString& fileName, bool deleteData)
 {
-    auto it = m_fileInformationByName.find(fileName);
+    auto it = m_fileInformationByName.find(keyFromFileName(fileName));
     if (it == m_fileInformationByName.end())
         return ResultCode::fileDoesNotExist;
 
-    const auto path = filePath(it->name);
-
-    if (deleteData && QFile::exists(path))
+    if (deleteData && QFile::exists(it->fullFilePath))
     {
-        if (!QFile::remove(path))
+        if (!QFile::remove(it->fullFilePath))
             return ResultCode::ioError;
     }
 
-    const auto metadataFileName = this->metadataFileName(path);
+    const auto metadataFileName = this->metadataFileName(it->fullFilePath);
     if (QFile::exists(metadataFileName))
     {
         if (!QFile::remove(metadataFileName))
             return ResultCode::ioError;
     }
 
-    QDir dir = QFileInfo(path).absoluteDir();
-    while (dir != m_downloadsDirectory)
+    // We won't delete custom user directories.
+    if (it->absoluteDirectoryPath.contains(m_downloadsDirectory.absolutePath()))
     {
-        if (dir.entryList(QDir::AllEntries | QDir::NoDotAndDotDot).isEmpty())
+        QDir dir = QFileInfo(it->absoluteDirectoryPath).absoluteDir();
+        while (dir != m_downloadsDirectory)
         {
-            const auto name = dir.dirName();
+            if (dir.entryList(QDir::AllEntries | QDir::NoDotAndDotDot).isEmpty())
+            {
+                const auto name = dir.dirName();
 
-            if (!dir.cdUp())
-                break;
+                if (!dir.cdUp())
+                    break;
 
-            if (!dir.rmdir(name))
+                if (!dir.rmdir(name))
+                    break;
+            }
+            else
+            {
                 break;
-        }
-        else
-        {
-            break;
+            }
         }
     }
 
@@ -453,7 +475,7 @@ QVector<QByteArray> Storage::getChunkChecksums(const QString& fileName)
 {
     QnMutexLocker lock(&m_mutex);
 
-    const auto& fileInfo = fileMetadata(fileName);
+    const auto& fileInfo = fileMetadata(keyFromFileName(fileName));
     if (!fileInfo.isValid())
         return QVector<QByteArray>();
 
@@ -465,7 +487,7 @@ ResultCode Storage::setChunkChecksums(
 {
     QnMutexLocker lock(&m_mutex);
 
-    auto it = m_fileInformationByName.find(fileName);
+    auto it = m_fileInformationByName.find(keyFromFileName(fileName));
     if (it == m_fileInformationByName.end())
         return ResultCode::fileDoesNotExist;
 
@@ -494,7 +516,7 @@ ResultCode Storage::setChunkChecksums(
     if (it->size < 0)
         return ResultCode::ok;
 
-    auto actualChecksums = calculateChecksums(filePath(fileName), it->chunkSize);
+    auto actualChecksums = calculateChecksums(it->fullFilePath, it->chunkSize);
     if (actualChecksums.size() != it->downloadedChunks.size())
         actualChecksums.resize(it->downloadedChunks.size());
 
@@ -520,7 +542,7 @@ void Storage::cleanupExpiredFiles()
     for (const FileMetadata& data: m_fileInformationByName)
     {
         if (data.ttl > 0 && data.touchTime + data.ttl <= currentTime)
-            expiredFiles.insert(data.name);
+            expiredFiles.insert(data.fullFilePath);
     }
 
     QSet<QString> successfullyDeletedFiles;
@@ -577,7 +599,7 @@ void Storage::findDownloadsRecursively(const QDir& dir)
                 const auto& name = m_downloadsDirectory.relativeFilePath(fileName);
                 {
                     NX_MUTEX_LOCKER lock(&m_mutex);
-                    if (m_fileInformationByName.contains(name))
+                    if (m_fileInformationByName.contains(keyFromFileName(name)))
                         continue;
                 }
                 loadDownload(fileName);
@@ -650,7 +672,7 @@ QVector<QByteArray> Storage::calculateChecksums(const QString& filePath, qint64 
 
 bool Storage::saveMetadata(const FileMetadata& fileInformation)
 {
-    const auto fileName = metadataFileName(filePath(fileInformation.name));
+    const auto fileName = metadataFileName(fileInformation.fullFilePath);
 
     QFile file(fileName);
     if (!file.open(QFile::WriteOnly))
@@ -733,9 +755,7 @@ void Storage::checkDownloadCompleted(FileMetadata& fileInfo)
             return;
     }
 
-    const auto path = filePath(fileInfo.name);
-
-    const auto md5 = calculateMd5(path);
+    const auto md5 = calculateMd5(fileInfo.fullFilePath);
     if (md5 != fileInfo.md5)
     {
         fileInfo.status = FileInformation::Status::corrupted;
@@ -743,7 +763,7 @@ void Storage::checkDownloadCompleted(FileMetadata& fileInfo)
     }
 
     fileInfo.status = FileInformation::Status::downloaded;
-    fileInfo.chunkChecksums = calculateChecksums(path, fileInfo.chunkSize);
+    fileInfo.chunkChecksums = calculateChecksums(fileInfo.fullFilePath, fileInfo.chunkSize);
 }
 
 ResultCode Storage::reserveSpace(const QString& fileName, const qint64 size)
